Update Products/CMFDeployment/DeploymentPolicy.py
Commented self.exportNotify(), so that by default, site Managers are not emailed a notification that deployment has completed.  IF Mailhost isn't setup, an error is thrown by this, which prevents deployment log from being written.  Nice option to re-enable, if a Mailhost is setup and the deployment is on a large site and runs for a long time.